% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibutils.R
\name{combn_char}
\alias{combn_char}
\title{Generate all combinations of the elements of a character vector}
\usage{
combn_char(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a nested list. A list whose elements are lists containing the 
character vectors with the combinations of their elements.
}
\description{
Generate all combinations of the elements of a character vector
}
\examples{
combn_char(c("gender", "sector"))
combn_char(c("gender", "sector", "education"))
}
