% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{crit_non_ann}
\alias{crit_non_ann}
\title{Get a normal distribution critical values but without annotation}
\usage{
crit_non_ann(df, level, mu, sigma, q, ci)
}
\arguments{
\item{df}{dataframe}

\item{level}{confidence level}

\item{mu}{population mean}

\item{sigma}{population standard deviation}

\item{q}{quantiles}

\item{ci}{confidence bounds}
}
\value{
a plot displaying critical values
}
\description{
Get a normal distribution critical values but without annotation
}
