% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binom.R
\name{critical_values_binom}
\alias{critical_values_binom}
\title{Make a vector of quantiles for the critical values of a binomial distribution}
\usage{
critical_values_binom(level, n, p)
}
\arguments{
\item{level}{confidence level}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
A vector of length 2 containing the upper and lower critical values
}
\description{
Make a vector of quantiles for the critical values of a binomial distribution
}
