% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process.R
\name{critical_values_exp}
\alias{critical_values_exp}
\title{Make a vector of quantiles for the critical values of a exponential distribution}
\usage{
critical_values_exp(level, rate)
}
\arguments{
\item{level}{confidence level}

\item{rate}{rate of occurence}
}
\value{
A vector of length 2 containing the upper and lower critical values
}
\description{
Make a vector of quantiles for the critical values of a exponential distribution
}
