% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process.R
\name{critical_values_gamma}
\alias{critical_values_gamma}
\title{Make a vector of quantiles for the critical values of a gamma distribution}
\usage{
critical_values_gamma(level, rate, shape)
}
\arguments{
\item{level}{confidence level}

\item{rate}{rate of occurence}

\item{shape}{shape parameter}
}
\value{
A vector of length 2 containing the upper and lower critical values
}
\description{
Make a vector of quantiles for the critical values of a gamma distribution
}
