% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{critical_values_norm}
\alias{critical_values_norm}
\title{Make a vector of quantiles for the critical values of a normal distribution}
\usage{
critical_values_norm(level, mu, sigma)
}
\arguments{
\item{level}{confidence level}

\item{mu}{population mean}

\item{sigma}{population standard deviation}
}
\value{
A vector of length 2 containing the upper and lower critical values
}
\description{
Make a vector of quantiles for the critical values of a normal distribution
}
