% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrr-package.R
\docType{package}
\name{distrr-package}
\alias{distrr}
\alias{distrr-package}
\title{distrr: Estimate and Manage Empirical Distributions}
\description{
Tools to estimate and manage empirical distributions, which should work with survey data. One of the main features is the possibility to create data cubes of estimated statistics, that include all the combinations of the variables of interest (see for example functions dcc5() and dcc6()).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gibonet.github.io/distrr}
  \item \url{https://github.com/gibonet/distrr}
  \item \url{https://gibonet.github.io/distrr/}
}

}
\author{
\strong{Maintainer}: Sandro Burri \email{gibo.gaf@gmail.com}

}
\keyword{internal}
