% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrr.R
\docType{package}
\name{distrr}
\alias{distrr}
\alias{distrr-package}
\title{Tools to estimate and manage empirical distributions}
\description{
Tools to estimate and manage empirical distributions,
which should work with survey data. One of the main features is the 
possibility to create data cubes of estimated statistics, that include
all the combinations of the variables of interest (see for example functions
'dcc' and 'dcc2').
}
