% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibutils.R
\name{extract_unique}
\alias{extract_unique}
\alias{extract_unique2}
\alias{extract_unique3}
\alias{extract_unique4}
\alias{extract_unique5}
\title{Functions to be used in conjunction with 'dcc' family}
\usage{
extract_unique(df)

extract_unique2(df)

extract_unique3(df)

extract_unique4(df)

extract_unique5(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a list whose elements are character vectors of the 
    unique values of each column
}
\description{
Functions to be used in conjunction with 'dcc' family
}
\examples{
data("invented_wages")
tmp <- extract_unique(df = invented_wages[, c("gender", "sector")])
tmp
str(tmp)
}
