% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fhat_conditional.R
\name{Fhat_df_}
\alias{Fhat_df_}
\title{Weighted empirical cumulative distribution function (data frame version)}
\usage{
Fhat_df_(.data, x, weights)
}
\arguments{
\item{.data}{a data frame}

\item{x}{name of the numeric column (as character)}

\item{weights}{name of the weight column (as character)}
}
\value{
a data frame with columns: x, wcum and Fhat
}
\description{
Weighted empirical cumulative distribution function (data frame version)
}
\examples{
data(invented_wages)
Fhat_df_(invented_wages, "wage", "sample_weights")
}
