% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binom.R
\name{get_annotation_location}
\alias{get_annotation_location}
\title{Get x-coordinate for annotation location on plot}
\usage{
get_annotation_location(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a value
}
\description{
Get x-coordinate for annotation location on plot
}
