% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binom.R
\name{get_ci_binom}
\alias{get_ci_binom}
\title{Make a vector of confidence interval bounds}
\usage{
get_ci_binom(q, n, p)
}
\arguments{
\item{q}{vector containing critical values}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
A vector of length 2 containing the upper and lower confidence bounds
}
\description{
Make a vector of confidence interval bounds
}
