% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{get_ci_norm}
\alias{get_ci_norm}
\title{Get the confidence bounds of normal distrbution}
\usage{
get_ci_norm(mu, sigma, q)
}
\arguments{
\item{mu}{population mean}

\item{sigma}{population standard deviation}

\item{q}{quantiles}
}
\value{
a vector of the lower and upper bounds
}
\description{
Get the confidence bounds of normal distrbution
}
