% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{make_normal_df}
\alias{make_normal_df}
\title{Make a data frame of quantiles and probabilities for a normal distribution}
\usage{
make_normal_df(mu, sigma)
}
\arguments{
\item{mu}{population mean}

\item{sigma}{population standard deviation}
}
\value{
A data frame containing quantiles and their corresponding probabilities for a given normal distribution
}
\description{
Make a data frame of quantiles and probabilities for a normal distribution
}
