% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{non_crit_non_ann}
\alias{non_crit_non_ann}
\title{Get a normal distribution without critical values and without annotation}
\usage{
non_crit_non_ann(df, level, mu, sigma, q, ci)
}
\arguments{
\item{df}{dataframe}

\item{level}{confidence level}

\item{mu}{population mean}

\item{sigma}{population standard deviation}

\item{q}{quantiles}

\item{ci}{confidence bounds}
}
\value{
a plot not displaying critical values and not displaying annotation
}
\description{
Get a normal distribution without critical values and without annotation
}
