% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binom.R
\name{non_crit_plot}
\alias{non_crit_plot}
\title{Create Binomial Plot with Critical Values}
\usage{
non_crit_plot(df, level, n, p, q, ci, annotate_x)
}
\arguments{
\item{df}{dataframe}

\item{level}{confidence level specified by user}

\item{n}{number of trials}

\item{p}{probability of success}

\item{q}{critical values}

\item{ci}{confidence bounds}

\item{annotate_x}{x-coordinate for the annotation of n, c, p, and the confidence interval}
}
\value{
a plot without critical values displayed
}
\description{
Create Binomial Plot with Critical Values
}
