% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{non_crit_plot_norm}
\alias{non_crit_plot_norm}
\title{Get a normal distribution without critical values}
\usage{
non_crit_plot_norm(df, level, mu, sigma, q, ci, annotate_x)
}
\arguments{
\item{df}{dataframe}

\item{level}{confidence level}

\item{mu}{population mean}

\item{sigma}{population standard deviation}

\item{q}{quantiles}

\item{ci}{confidence bounds}

\item{annotate_x}{x-coordinate for annotation}
}
\value{
a plot displaying critical values
}
\description{
Get a normal distribution without critical values
}
