% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binom.R
\name{plot_binom}
\alias{plot_binom}
\title{Plot the Binomial Distribution}
\usage{
plot_binom(n, p, level = 0.95, crit = T)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}

\item{level}{confidence level}

\item{crit}{(T/F) whether or not you want critical values displayed}
}
\value{
A plot of the binomial distribution with user-specified details
}
\description{
Plot the Binomial Distribution
}
