% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process.R
\name{plot_exponential}
\alias{plot_exponential}
\title{Display Exponential plot}
\usage{
plot_exponential(rate, level)
}
\arguments{
\item{rate}{rate of occurrence}

\item{level}{confidence level}
}
\value{
an exponential plot
}
\description{
Display Exponential plot
}
