% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Display Gamma Plot}
\usage{
plot_gamma(rate, shape, level)
}
\arguments{
\item{rate}{rate of occurrence}
}
\value{
a gamma plot
}
\description{
Display Gamma Plot
}
