% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{plot_normal}
\alias{plot_normal}
\title{Plot the Normal Distribution}
\usage{
plot_normal(mu, sigma, level = 0.95)
}
\arguments{
\item{mu}{population mean}

\item{sigma}{population standard deviation}

\item{level}{confidence level}
}
\value{
A plot of the normal distribution with user-specified details
}
\description{
Plot the Normal Distribution
}
