% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process.R
\name{plot_poisson}
\alias{plot_poisson}
\title{Display a Poisson plot}
\usage{
plot_poisson(rate, level)
}
\arguments{
\item{rate}{rate of occurrence}

\item{level}{confidence level}
}
\value{
a poisson plot
}
\description{
Display a Poisson plot
}
