% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process.R
\name{poisson_process}
\alias{poisson_process}
\title{Specify a plot of any part of the poisson process}
\usage{
poisson_process(dist, rate, shape = 1, level = 0.95)
}
\arguments{
\item{dist}{the specified distribution(Poisson, Exponential, Gamma)}

\item{rate}{rate paramater AKA rate at which a single event occurs}

\item{shape}{shape parameter (neccessary for Gamma) AKA number of events waiting to occur}

\item{level}{confidence level}
}
\value{
a plot of specified distribution
}
\description{
Specify a plot of any part of the poisson process
}
