% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{return_list}
\alias{return_list}
\title{Build result list at the end of plot_normal()}
\usage{
return_list(
  df,
  level,
  q,
  ci,
  mu,
  sigma,
  crit_plot,
  crit_non_ann,
  non_crit_plot,
  non_crit_non_ann
)
}
\arguments{
\item{df}{data frame}

\item{level}{confidence level}

\item{q}{quantiles}

\item{ci}{confidence bounds}

\item{mu}{population mean}

\item{sigma}{population standard deviation}

\item{crit_plot}{critical value plot with annotation}

\item{crit_non_ann}{critical value plot without annotation}

\item{non_crit_plot}{non-critical value plot with annotation}

\item{non_crit_non_ann}{non-critical value plot without annotation}
}
\value{
a list
}
\description{
Build result list at the end of plot_normal()
}
